enabled task deregistration for testing

+++ EOSS/aws/EvalQueue.py
@@ -1,6 +1,6 @@
 import boto3
 
-from EOSS.aws.utils import dev_client, prod_client
+from EOSS.aws.utils import dev_client, prod_client, user_input
 
 
 class EvalQueue:
@@ -52,6 +52,9 @@ class EvalQueue:
     def delete_all_eval_queues(self):
         urls = self.get_all_eval_queue_urls()
         print('---> QUEUES TO DELETE', urls)
+        if not user_input('\n ---> The queues above are going to be deleted, would you like to continue? (yes/no):  '):
+            exit(0)
+
         for url in urls:
             self.delete_queue(url)
